// gesit92-webapps-development/gerenciador-app/src/app/admin/admin-enum/config.enum.ts
export const enum ConfiguracaoEnum {
    
 URL_BASE_PROD  = "https://capacidade.servicos.bb.com.br:3001/capacidade",
 URL_BASE_DESENV = "https://capacidade.servicos.bb.com.br:3001/capacidade",
 URL_AVATAR_HUMANOGRAMA = "https://humanograma.intranet.bb.com.br/avatar/",
 IMG_DEFULT = "assets/img/default.jpg",
 COOKIE_NAME = "uid"

}   